Move thunks to their own files; Start updating ActivityList to v0.2
