+++ app/scenes/Activity/index.js
@@ -2,7 +2,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Actions } from 'react-native-router-flux';
-import { setAnswer, completeResponse } from '../../state/responses/responses.actions';
+import { setAnswer } from '../../state/responses/responses.actions';
+import { completeResponse } from '../../state/responses/responses.thunks';
 import { currentActivitySelector, currentResponsesSelector } from '../../state/responses/responses.selectors';
 import Screen from '../../components/screen';
 import ActivityComponent from './ActivityComponent';
