+++ app/state/app/app.reducer.test.js
@@ -4,12 +4,6 @@ import {
   resetApiHost,
 } from './app.actions';
 
-jest.mock('react-native-device-info', () => { });
-jest.mock('react-native-fetch-blob', () => { });
-jest.mock('react-native-push-notification', () => { });
-jest.mock('react-native-router-flux', () => { });
-jest.mock('native-base', () => { });
-
 test('it has an initial state', () => {
   expect(appReducer(undefined, { type: 'foo' })).toEqual(initialState);
 });
