+++ app/state/responses/responses.reducers.test.js
@@ -12,12 +12,6 @@ import {
   shiftUploadQueue,
 } from './responses.actions';
 
-jest.mock('react-native-device-info', () => { });
-jest.mock('react-native-fetch-blob', () => { });
-jest.mock('react-native-push-notification', () => { });
-jest.mock('react-native-router-flux', () => { });
-jest.mock('native-base', () => { });
-
 test('it has an initial state', () => {
   expect(responsesReducer(undefined, { type: 'foo' })).toEqual(initialState);
 });
